Extract composer UI motion helpers
